# Faster bundler for node & fix npmignore