--- a/js/package.json
+++ b/js/package.json
@@ -36,7 +36,6 @@
     "ws": "^8.16.0"
   },
   "devDependencies": {
-    "@types/commander": "^2.12.2",
     "@types/expect-puppeteer": "^5.0.6",
     "@types/jest": "^29.5.12",
     "@types/jest-environment-puppeteer": "^5.0.6",