--- a/js/tsup.config.ts
+++ b/js/tsup.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+  name: 'buttplug-js',
+  entry: ['src/index.ts'],
+  target: ['es6'],
+  format: ['cjs', 'esm'],
+  outDir: 'dist/main',
+  platform: 'node',
+  clean: true,
+  sourcemap: true,
+  dts: true,
+});