--- a/js/tsup.config.ts
+++ b/js/tsup.config.ts
@@ -10,4 +10,5 @@ export default defineConfig({
   clean: true,
   sourcemap: true,
   dts: true,
+  minify: true,
 });